Relocate fast render method to Renderable concern
